cambio de tipo de dato calificacion

+++ src/main/java/cl/ecommerce/ecommerce/dto/PublicacionDTO.java
@@ -41,6 +41,6 @@ public class PublicacionDTO {
     private Date fecha;
 
     @Column(nullable = false)
-    private Integer calificacion;
+    private Long calificacion;
 
 }
